pyproject.toml
[project]
name = "archi-n-tier-admin-flask"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.0",
    "python-dotenv>=1.1.0",
    "sqlalchemy>=2.0.40",
]

[dependency-groups]
dev = [
    "ruff>=0.11.7",
]
